flask_app/requirements.txt
alembic==1.6.2
blinker==1.4
click==7.1.2
decorator==5.0.7
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Bootstraps==3.3.7.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flask-wtf-alchemy-utils==0.1.0
greenlet==1.1.0
idna==3.1
infinity==1.5
intervals==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.15
SQLAlchemy-Utils==0.37.2
validators==0.18.2
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.5
